requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
sqlparse==0.4.4